[Feature] Move characters list screen to characters.list package
