+++ app/src/main/java/com/vadymmy/ricktionary/ui/characters/list/CharactersContract.kt
@@ -1,6 +1,6 @@
-package com.vadymmy.ricktionary.ui.characters
+package com.vadymmy.ricktionary.ui.characters.list
 
-import com.vadymmy.ricktionary.ui.characters.model.CharacterItemUiModel
+import com.vadymmy.ricktionary.ui.characters.list.model.CharacterItemUiModel
 
 data class CharactersUiState(
     val isLoading: Boolean = true,
