+++ app/src/main/java/com/vadymmy/ricktionary/ui/characters/list/model/CharacterItemUiModel.kt
@@ -1,4 +1,4 @@
-package com.vadymmy.ricktionary.ui.characters.model
+package com.vadymmy.ricktionary.ui.characters.list.model
 
 data class CharacterItemUiModel(
     val id: Int,
